+++ packages/story-editor/src/components/library/test/text/textTab.js
@@ -25,7 +25,7 @@ import { curatedFontNames } from '@web-stories-wp/fonts';
  */
 import { renderWithTheme } from '../../../../testUtils';
 import FontContext from '../../../../app/font/context';
-import fontsListResponse from '../../../form/advancedDropDown/test/fontsResponse.json';
+import fontsListResponse from '../../../panels/design/textStyle/test/fontsResponse.json';
 import TextIcon from '../../panes/text/textIcon';
 import { DEFAULT_PRESET } from '../../panes/text/textPresets';
 import useLibrary from '../../useLibrary';
